change product id to long

+++ src/main/java/com/ecommerce/repositories/impl/ProductRepositoryImpl.java
@@ -4,7 +4,7 @@ package com.ecommerce.repositories.impl;
 import com.ecommerce.repositories.ProductRepository;
 import com.ecommerce.repositories.entites.ProductEntity;
 
-public class ProductRepositoryImpl extends RepositoryImpl<ProductEntity,Integer> implements ProductRepository {
+public class ProductRepositoryImpl extends RepositoryImpl<ProductEntity,Long> implements ProductRepository {
     private static final ProductRepositoryImpl INSTANCE =new ProductRepositoryImpl();
     private ProductRepositoryImpl(){
